+++ src/features/template_resolvers/mod.rs
@@ -2,4 +2,11 @@
 mod vite_template_resolver;
 
 #[cfg(feature = "vite-template-resolver")]
+#[allow(deprecated)]
 pub use vite_template_resolver::ViteTemplateResolver;
+
+#[cfg(feature = "vite-hbs-template-resolver")]
+mod vite_hbs_template_resolver;
+
+#[cfg(feature = "vite-hbs-template-resolver")]
+pub use vite_hbs_template_resolver::ViteHBSTemplateResolver;